bugfix: `vsnTag` replaced with `vsntag`
Somewhere earlier, @sih and I had a discussion on the case used for tags. Not being able to find where the decision was documented, I looked through the specifications and found that there was only one occurrence of (scope|vsn|group)Tag in the specifications, which this commit has fixed.

There are, however mentions of `groupTag` or `groupTags` remaining, but they are (mostly) in the front matter. In another commit, they will also all be renamed into `grouptag` resp. `grouptags` (all lower case).